refactor: move compat.h into compat/
