+++ src/test/fuzz/util.h
@@ -8,7 +8,7 @@
 #include <arith_uint256.h>
 #include <chainparamsbase.h>
 #include <coins.h>
-#include <compat.h>
+#include <compat/compat.h>
 #include <consensus/amount.h>
 #include <consensus/consensus.h>
 #include <merkleblock.h>
